Cast to int to avoid signed/unsignd mismatch

+++ test/core/image/test_fixture.hpp
@@ -28,10 +28,12 @@ using image_types =
                gil::rgba8_image_t, gil::rgba16_image_t, gil::rgba32_image_t>;
 
 template <typename T> struct consecutive_value {
-  consecutive_value(T start) : current_(start) { BOOST_TEST(current_ >= 0); }
+  consecutive_value(T start) : current_(start) {
+    BOOST_TEST(static_cast<int>(current_) >= 0);
+  }
 
   T operator()() {
-    BOOST_ASSERT(current_ + 1 > 0);
+    BOOST_ASSERT(static_cast<int>(current_) + 1 > 0);
     return current_++;
   }
 
@@ -40,7 +42,7 @@ template <typename T> struct consecutive_value {
 
 template <typename T> struct reverse_consecutive_value {
   reverse_consecutive_value(T start) : current_(start) {
-    BOOST_ASSERT(current_ > 0);
+    BOOST_ASSERT(static_cast<int>(current_) > 0);
   }
 
   T operator()() {
